[misc.roundTools] New module; move otRound() here
Code relying on old place (fixedTools.otRound) still works.
