+++ Lib/fontTools/misc/fixedTools.py
@@ -18,6 +18,7 @@ functions for converting between fixed-point, float and string representations.
 """
 
 from fontTools.misc.py23 import *
+from .roundTools import otRound
 import math
 import logging
 
@@ -25,7 +26,6 @@ log = logging.getLogger(__name__)
 
 __all__ = [
 	"MAX_F2DOT14",
-	"otRound",
 	"fixedToFloat",
 	"floatToFixed",
 	"floatToFixedToFloat",
@@ -39,30 +39,6 @@ __all__ = [
 
 
 MAX_F2DOT14 = 0x7FFF / (1 << 14)
-
-
-def otRound(value):
-	"""Round float value to nearest integer towards ``+Infinity``.
-
-	The OpenType spec (in the section on `"normalization" of OpenType Font Variations <https://docs.microsoft.com/en-us/typography/opentype/spec/otvaroverview#coordinate-scales-and-normalization>`_)
-	defines the required method for converting floating point values to
-	fixed-point. In particular it specifies the following rounding strategy:
-
-		for fractional values of 0.5 and higher, take the next higher integer;
-		for other fractional values, truncate.
-
-	This function rounds the floating-point value according to this strategy
-	in preparation for conversion to fixed-point.
-
-	Args:
-		value (float): The input floating-point value.
-
-	Returns
-		float: The rounded value.
-	"""
-	# See this thread for how we ended up with this implementation:
-	# https://github.com/fonttools/fonttools/issues/1248#issuecomment-383198166
-	return int(math.floor(value + 0.5))
 
 
 def fixedToFloat(value, precisionBits):
