+++ Lib/fontTools/ttLib/tables/otTables.py
@@ -9,7 +9,7 @@ from enum import IntEnum
 import itertools
 from collections import namedtuple
 from fontTools.misc.py23 import *
-from fontTools.misc.fixedTools import otRound
+from fontTools.misc.roundTools import otRound
 from fontTools.misc.textTools import pad, safeEval
 from .otBase import (
 	BaseTable, FormatSwitchingBaseTable, ValueRecord, CountReference,
